+++ src/lib/validate/types/index.ts
@@ -1,0 +1,13 @@
+// export from here in case you import from validatees/dist/lib/types
+// types
+export { default as isFalsy } from "./isFalsy";
+export { default as isFalsyExtended } from "./isFalsyExtended";
+export { default as isTruthy } from "./isTruthy";
+export { default as isTruthyExtended } from "./isTruthyExtended";
+export { default as isNullish } from "./isNullish";
+export { default as isString } from "./isString";
+export { default as isNumber } from "./isNumber";
+export { default as isBoolean } from "./isBoolean";
+export { default as isExtendable } from "./isExtendable";
+export { default as isRegExp } from "./isRegExp";
+export { default as isDate } from "./isDate";