+++ tests/types/isTruthy.test.ts
@@ -1,4 +1,4 @@
-import isTruthy from "../../src/lib/types/isTruthy";
+import isTruthy from "../../src/lib/validate/types/isTruthy";
 describe("isTruthy", () => {
   it("should return true for 1", () => {
     expect(isTruthy(1)).toBe(true);